requirements.txt
# Core ML and AI Libraries (Updated Phase 2 - CUDA Enabled)
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
transformers==4.56.2
accelerate==1.10.1
bitsandbytes==0.47.0
peft==0.6.0
huggingface-hub==0.35.1
safetensors==0.6.2

# Analytics and Visualization
matplotlib==3.8.0
pandas==2.1.0
seaborn==0.12.0

# API and Web Framework (Phase 2 Enhanced)
fastapi==0.117.1
uvicorn==0.37.0
websockets==12.0
python-socketio==5.11.0
python-multipart==0.0.6
pydantic==2.11.9
starlette==0.48.0

# OpenAI Integration
openai==1.3.6
whisper-openai==20231117

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
speech-recognition==3.10.0

# Data Processing
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Visualization (for C# integration)
matplotlib==3.7.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0